Statemachine and model added. Jupyter notebook examples created (#1)
State machine and model added, the basic skeleton created for enabling a command-able Scheduler. Currently only template data is used. The model needs to be worked on to include methods which allow the pulling of real targets from the Driver. As well as methods on the model which can set settings for the various objects the scheduling algorithm uses for computational work.