Implement `start` command on the Scheduler CSC.  (#18)
Implement `start` command on the Scheduler CSC. The start command takes some time to load and configure the models, driver and the scheduling algorithm. While it is configuring it must keep its responsiveness. To allow that, we use a `ThreadPoolExecutor` to run the configuration task. The models now need to be updated to contain a `parameters` object that host the configuration.

Adds unit test for the SchedulerCSC. It will bring the CSC from OFFLINE to ENABLE and back. Still need to add some checks for responsiveness when it is starting (checked manually and it is working, but need unit test).

Updated scheduler initialization script to initialize the CSC. The script may need some extra polishing later.

Disable unit tests for deprecated Model class. This will be removed later.

If log_path does not exists, create it.