In `scheduler_csc`, refactoring  `advance_target_loop`.
Extract method to add targets to the queue into a new method `queue_targets`.
This method will switch the detailed state to `QUEUEING_TARGET` and execute the operations previously done inline in `advance_target_loop`.
If the method fails to producer targets and fails to find a target in the future, it raises a new exception `FailedToQueueTargetsError`.