In `scheduler_csc`, continue refactor of `advance_target_loop`, change how `generate_target_queue` and `compute_predicted_schedule` runs.
Instead of calling them sequentially, which causes a delay in producing targets, call `generate_target_queue` first and then, while waiting for targets to execute, call `compute_predicted_schedule`.
This new way improves over the previous implementation but still leave room for conditions where `compute_predicted_schedule` can take longer to execute than the targets to be observed and then lead to additional delay.
Nevertheless, conditions like this were only observed with the AuxTel configuration, which is a lot heavier than the one for the main telescope, and because there was a bug in the predictive loop which would continue to compute targets until the queue was filled up.