Add new submodule ``utils/scheduled_targets_info.py`` defining a dataclass `ScheduledTargetsinfo`` wich holds information about scheduled targets.
This is going to be used by the ``check_scheduled_targets`` method in the (new) ``Model`` class to return information about targets that were verified.
This is part of a refactor of the SchedulerCSC code which extracts into the ``Model`` class some of the functionality related to the control loop operation, reducing the ammount of code in the CSC and improving separation of concerns.