Add new submodule ``model.py`` which implements the ``Model`` class.
This class holds the bulk of the logic behind the control loop operations.
It contains extracts the following information from the SchedulerCSC class:

* telemetry_stream_handler: Handle retrieval of telemetry from the EFD.
* models: Dictionary with the available models.
* raw_telemetry: Dictionary with all the raw telemetry values.
* script_info: Dictionary with information about queued scripts.
* driver: Instance of the scheduler driver, that encapsulates the mechanism to drive the scheduling algorithm.

In addition, the ``Model`` class now extracts from the CSC class all the logic to manage these resources, such that the CSC can call the model to execute the required operations.
The idea is to improve the separation of concerns, impriving the maintainability  of the package, and reducing bloat in the CSC code.