In ``scheduler_csc.py``, implement limit on the number of scripts the CSC will send to the queue.
This is done by using a semaphore with a capacity of 10 (non-configurable).
Each time the CSC adds a script to the Queue it acquires the semaphore, then creates a background task that also acquires the semaphore.
The background task will monitor the state of the script and only finishes when the script reaches a final state, thus releasing the semaphore.
The background task sets the value of a future to either be the final state of the script or an exception, if the final state indicates a failure.

Before adding new scripts to the Queue, the CSC checks the state of the previous finished scripts.
If any of them failed the procedure is interrupted and any script previously sent to the queue is removed.
This also happens if a user adds a Script to the queue in the middle of an executing block.

This also fixes a previous bug with the queue cleanup method that would cause the Scheduler to remove a failed script from the top of the queue.
The cleanup method now checks and only removes scripts that have not failed.