In driver/driver_target.py, add filter_name to the script configuration.
There is an important distinction between the band_name and the filter_name. While band_name is the common name used to classify a filter (e.g. ugrizy) the filter name is the name of that particular element (SDSSr_65mm or r_07_xyz). Usually the instrument will use filter name to select an element but other operations will refer to the band name.